refactor: extract audit purge to background.go for maintainability
Move startAuditPurge, runAuditPurge, and auditPurgeInterval from
main.go to background.go (same package). Move the corresponding
tests to background_test.go. This reduces main.go from 610 to 539
lines and groups background job code into a self-contained file
that's easier to find and extend.
